[#63]Chore(게시글): jpa 쿼리 때문에 commentCount 타입을 int 에서 Long 으로 변환

+++ src/main/java/today/sesac/shoutify/comment/repository/CommentRepository.java
@@ -32,5 +32,5 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
      * @param postId 게시글 Id
      */
     @Query("SELECT COUNT(c) FROM Comment c WHERE c.post.id = :postId")
-    int countByPostId(@Param("postId") Long postId);
+    Long countByPostId(@Param("postId") Long postId);
 }
